% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{by_strata_DocTopic}
\alias{by_strata_DocTopic}
\title{Estimate Document-Topic distribution by strata}
\usage{
by_strata_DocTopic(x, by_name, by_values, burn_in = NULL,
  parallel = TRUE, mc.cores = NULL)
}
\arguments{
\item{x}{the output from a keyATM model (see \code{keyATM()})}

\item{by}{a vector whose length is the number of documents}
}
\value{
strata_topicword object (a list)
}
\description{
Estimate Document-Topic distribution by strata
}
