% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{calc_phi_by}
\alias{calc_phi_by}
\title{Estimate Subsetted Topic-Word distribution}
\usage{
calc_phi_by(x, keyATM_docs, by)
}
\arguments{
\item{x}{the output from a keyATM model (see \code{keyATM()})}

\item{keyATM_docs}{(see \code{keyATM_read()})}

\item{by}{a vector whose length is the number of documents}
}
\description{
Estimate Subsetted Topic-Word distribution
}
