% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topicdict-package.R
\docType{data}
\name{corpus_bara_para}
\alias{corpus_bara_para}
\title{Bara et al's Abortion Debate Corpus}
\format{A quanteda corpus
\describe{
  \item{speaker}{Name of speaker}
  \item{vote}{Whether they voted yes, no, or abstained}
  ...
}}
\source{
Bara et al. 2007 Swiss Political Science Review
}
\usage{
corpus_bara_para
}
\description{
A corpus of speeches from an Abortion debate in the UK House
of Commons, studied by Bara et al. 2007.  The unit of analysis
is the paragraph entry (406). Speaker and final vote metadata is included.
}
\keyword{datasets}
