% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topicdict-package.R
\docType{data}
\name{corpus_uk_platforms}
\alias{corpus_uk_platforms}
\title{Post-War UK Party Platforms}
\format{A quanteda corpus
\describe{
  \item{date}{ Election year}
  \item{party}{ Party abbreviation}
}}
\source{
Ken Benoit
}
\usage{
corpus_uk_platforms
}
\description{
A corpus of party platforms since 1945 in the United Kingdom.
}
\keyword{datasets}
