% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{covariates_get}
\alias{covariates_get}
\title{Return covariates used in the iteration}
\usage{
covariates_get(x)
}
\arguments{
\item{x}{the output from a keyATM model (see \code{keyATM()})}
}
\description{
Return covariates used in the iteration
}
