% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{covariates_info}
\alias{covariates_info}
\title{Show covariates information}
\usage{
covariates_info(x)
}
\arguments{
\item{x}{the output from a keyATM model (see \code{keyATM()})}
}
\description{
Show covariates information
}
