% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topicdict_cpp.R
\name{create_seed_file}
\alias{create_seed_file}
\title{Create Seed file}
\usage{
create_seed_file(seed_words, num_regular_topic = 0,
  save_file_name = "seedwords")
}
\arguments{
\item{num_regular_topic}{a number of regular topic}

\item{seeds_words}{a vector of seed words.}
}
\description{
Create Seed file
}
