% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata_call.R
\name{create_sim_data}
\alias{create_sim_data}
\title{Create Simulation Data}
\usage{
create_sim_data(saveDir, D = 200, K = 10, TotalV = 8000, alpha = 0.1,
  beta_r = 0.1, beta_s = 0.1, p = NULL, gamma1 = NULL, gamma2 = NULL,
  lambda = 300, seeds_len = 5, prop_seeds = NULL, rand_seed = 123)
}
\arguments{
\item{saveDir}{data storage (set full path)}

\item{D}{number of documents}

\item{K}{number of topics}

\item{TotalV}{maximum number of unique words}

\item{alpha}{parameter of dirichlet of theta}

\item{beta_r}{parameter of dirichlet of phi_regular}

\item{beta_s}{parameter of dirichlet of phi_seed}

\item{p}{probability of using seed (set either gamma or p)}

\item{gamma1}{parameter of beta (set either gamma or p)}

\item{gamma2}{parameter of beta (set either gamma or p)}

\item{lambda}{parameter of poisson}

\item{seeds_len}{number of seeds words per each topic}

\item{prop_seeds}{proportion of seed words}
}
\description{
Create Simulation Data
}
\examples{
\dontrun{create_sim_data(saveDir="/Users/Shusei/Desktop/temp/SimulationData/", D=200, K=10, TotalV=8000, alpha=0.1, beta_r=0.1, beta_s=0.1, gamma1=1, gamma2=20, lambda=300, seeds_len=5)}
\dontrun{create_sim_data(saveDir="/Users/Shusei/Desktop/temp/SimulationData/", D=200, K=10, TotalV=8000, alpha=0.1, beta_r=0.1, beta_s=0.1, p=0.3, lambda=300, seeds_len=5)}

}
