% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ddirmnCpp}
\alias{ddirmnCpp}
\title{Dirichlet Multinomial Distribution}
\usage{
ddirmnCpp(Y, Lambda, X)
}
\arguments{
\item{Y}{outcome (a table of Z, n_dk)}

\item{Lambda}{current coefficients}

\item{X}{covariates}
}
\description{
Dirichlet Multinomial Distribution
}
\keyword{internal}
