% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{diagnosis_alpha}
\alias{diagnosis_alpha}
\title{Show a diagnosis plot of alpha}
\usage{
diagnosis_alpha(res, start = NULL, show_topic = NULL, true_vec = NULL,
  scale = "")
}
\arguments{
\item{start}{Slice iteration}

\item{show_topic}{a vector to specify topic indexes to show}

\item{true_vec}{a vector to visualize true values of alpha}

\item{x}{The posterior from a fitted model (see \code{posterior})}
}
\value{
ggplot2 object
}
\description{
Show a diagnosis plot of alpha
}
