% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{diagnosis_p}
\alias{diagnosis_p}
\title{Show a diagnosis plot of p}
\usage{
diagnosis_p(x, show_topic = NULL)
}
\arguments{
\item{x}{the output from a keyATM model (see \code{keyATM_output()})}

\item{show_topic}{A vector to indicate topics to visualize}
}
\value{
ggplot2 object
}
\description{
Show a diagnosis plot of p
}
