% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{estimate_coefficients}
\alias{estimate_coefficients}
\title{Regress theta on covariates}
\usage{
estimate_coefficients(x, covariates_data, covariates_formula = NULL,
  thinning = 5, burn_in = NULL, parallel = TRUE, mc.cores = NULL)
}
\arguments{
\item{x}{the output from a keyATM model (see \code{keyATM()})}

\item{covariates_data}{matrix, data.frame, or tibble}

\item{covariates_formula}{}
}
\value{
keyATM_coefficients object
}
\description{
Regress theta on covariates
}
