% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{explore}
\alias{explore}
\title{Explore Documents}
\usage{
explore(files, encoding = "UTF-8", lowercase = TRUE,
  remove_numbers = TRUE, remove_punct = TRUE, remove_symbols = TRUE,
  remove_separators = TRUE, remove_twitter = FALSE,
  remove_hyphens = FALSE, remove_url = TRUE, stem_language = NULL,
  stopwords = NULL)
}
\arguments{
\item{files}{names of each file to read}

\item{encoding}{File encoding (Default: whatever \code{quanteda} guesses)}

\item{lowercase}{whether to transform each token to lowercase letters}

\item{remove_numbers}{whether to remove numbers}

\item{remove_punct}{whether to remove punctuation}

\item{remove_symbols}{whether to remove non-alphanumerical symbols}

\item{remove_separators}{whether to remove tabs, spaces, newlines, etc.}

\item{remove_twitter}{whether to remove Twitter detritus}

\item{remove_hyphens}{whether to split hyphenated words}

\item{remove_url}{whether to remove URLs}

\item{stem_language}{if not NULL, the language to use for stemming}

\item{stopwords}{if not NULL, a character vector of words to remove,
e.g. \code{quanteda::stopwords("english")}}
}
\value{
A R4 objects \describe{
        \item{show_words(word, type="count", n_show=100, xaxis=F)}{ show a word distribution. \code{type} should be \code{count} or \code{proportion}.}
        \item{top_words(n_show=10)}{ show top words}
        }.
}
\description{
Explore the documents.
}
