% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{class}
\name{ExploreDocuments}
\alias{ExploreDocuments}
\alias{explore_class}
\title{Explore Documents Class}
\value{
A R4 objects \describe{
        \item{show_words(word, type="count", n_show=100, xaxis=F)}{show a word distribution. \code{type} should be \code{count} or \code{proportion}.}
        \item{top_words(n_show=10)}{show top words}
        }.
}
\description{
Explore Documents Class
}

\section{Fields}{

\itemize{
   \item data data in tidytext format
   \item uniquewords a vector of unique words
   \item totalwords number of words in the entire documents
}
}

\section{Contains}{

NULL
}

\section{Methods}{

 \itemize{
   \item initialize
   \item check_existence
   \item show_words
   \item top_words
 }
}

