% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyATM-package.R
\docType{package}
\name{keyATM-package}
\alias{keyATM-package}
\title{keyATM: Keyword Assisted Topic Models}
\description{
keyATM fits Keyword Assisted Topic Models
}
\seealso{
Useful links:
\itemize{
  \item \url{https://shusei-e.github.io/keyATM/}
  \item Report bugs at \url{https://github.com/Shusei-E/keyATM/issues}
}

}
\author{
\strong{Maintainer}: Shusei Eshima \email{shuseieshima@g.harvard.edu}

Authors:
\itemize{
  \item Tomoya Sasaki \email{tomoyas@mit.edu}
  \item Kosuke Imai \email{imai@harvard.edu}
}

Other contributors:
\itemize{
  \item William Lowe \email{wlowe@princeton.edu} [contributor]
}

}
