% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{keyATM_data_bills}
\alias{keyATM_data_bills}
\title{Bills data}
\format{A list with following objects:
\describe{
  \item{doc_dfm}{A \code{quanteda} dfm object of 140 documents}
  \item{cov}{}
  \item{keywords}{}
  \item{time_index}{}
}}
\source{
\url{http://}
}
\usage{
keyATM_data_bills
}
\description{
Bills data
}
\keyword{datasets}
