% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{keyATM_fit_cov}
\alias{keyATM_fit_cov}
\title{Run the Collapsed Gibbs sampler for the keyATM covariates (Dir-Multi)}
\usage{
keyATM_fit_cov(model, resume = FALSE)
}
\arguments{
\item{model}{A initialized model}

\item{resume}{resume or not}
}
\description{
Run the Collapsed Gibbs sampler for the keyATM covariates (Dir-Multi)
}
\keyword{internal}
