% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{keyATM_fit_HMM}
\alias{keyATM_fit_HMM}
\title{Run the Collapsed Gibbs sampler for the keyATM Dynamic}
\usage{
keyATM_fit_HMM(model, resume = FALSE)
}
\arguments{
\item{model}{A initialized model}

\item{resume}{resume or not}
}
\description{
Run the Collapsed Gibbs sampler for the keyATM Dynamic
}
\keyword{internal}
