% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{keyATM_idealpoint}
\alias{keyATM_idealpoint}
\title{Run the Collapsed Gibbs sampler for Ideal Point Estimation Model}
\usage{
keyATM_idealpoint(model, author_info, iter = 0L, output_iter = 10L)
}
\arguments{
\item{model}{A initialized model}

\item{author_info}{author information}

\item{iter}{Required number of iterations}

\item{output_per}{Show log-likelihood and perplexity per this number during the iteration}
}
\description{
Run the Collapsed Gibbs sampler for Ideal Point Estimation Model
}
