% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{keyATM_model}
\alias{keyATM_model}
\title{Initialize a keyATM model}
\usage{
keyATM_model(files = NULL, keywords = NULL, text_df = NULL,
  text_dfm = NULL, mode = "", extra_k = 1, covariates_data = NULL,
  covariates_formula = NULL, num_states = NULL, timestamps = NULL,
  alpha = 50/(length(keywords) + extra_k), beta = 0.01, beta_s = 0.1,
  options = list())
}
\description{
Initialize a keyATM model
}
