% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{keyATM_read}
\alias{keyATM_read}
\title{Read texts}
\usage{
keyATM_read(texts, encoding = "UTF-8", check = TRUE)
}
\arguments{
\item{texts}{Input. keyATM_read takes dfm, data.frame, \pkg{tibble} tbl_df, or a vector of file paths.}

\item{encoding}{A character. Only used when \code{texts} is a vector of file paths. Default is \code{UTF-8}.}

\item{check}{logical; If \code{TRUE}, check whether there are any problems in the structure of texts (e.g. extra spaces between texts). Default is \code{TRUE}.}
}
\value{
A keyATM_docs-class object. A list whose elements are splitted texts. The length of the list equals to the number of documents.
}
\description{
This function reads texts and creates a keyATM_docs object, which is a list of texts. The currently available sources are: 
\itemize{
   \item a quanteda dfm object (\strong{recommended})
   \item a data.frame (or a \pkg{tibble} tbl_df) with texts contained in a column named \code{text}
   \item a vector of file paths to text source files
   }
}
\examples{
\dontrun{
 # Use quanteda dfm (recommended)
 keyATM_docs <- keyATM_read(texts = quanteda_dfm) 
  
 # Use data.frame or tibble (texts should be stored in a column named `text`)
 keyATM_docs <- keyATM_read(texts = data_frame_object) 
 keyATM_docs <- keyATM_read(texts = tibble_object) 

 # Use a vector that stores full paths to the text files  
 files <- list.files(doc_folder, pattern = "*.txt", full.names = T) 
 keyATM_docs <- keyATM_read(texts = files) 
}

}
