% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{keyATM_run}
\alias{keyATM_run}
\title{Run a keyATM model}
\usage{
keyATM_run(texts, keywords, mode, extra_k, iteration = 1000,
  covariates_data = NULL, covariates_formula = ~. + 0,
  timestamps = NULL, options = list())
}
\arguments{
\item{texts}{Inputs. It can take quanteda dfm, data.frame, tibble, and a vector of characters.}

\item{keywords}{a quanteda dictionary or a list of character vectors}

\item{mode}{number of iteration}

\item{extra_k}{number of regular topics in addition to the keyword topics by
\code{keywords}}

\item{iteration}{number of iteration}

\item{covariates_data}{covariate}

\item{covariates_formula}{formula applied to covariate data}

\item{timestamps}{timestamps}

\item{options}{options are seed (default: 225), use_weights (default: 1),
 visualize_keywords (default: TRUE),
slice_shape (default: 1.2), and output_per (default: 10).}
}
\value{
A list
}
\description{
This function takes a text input, initialize and fit keyATM model.
}
