% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{keyATM_train_tot}
\alias{keyATM_train_tot}
\title{Run the Collapsed Gibbs sampler for the topics-over-time model}
\usage{
keyATM_train_tot(model, iter = 0L, output_per = 10L)
}
\arguments{
\item{model}{A initialized model}

\item{iter}{Required number of iterations}

\item{output_per}{Show log-likelihood and perplexity per this number during the iteration}
}
\description{
Run the Collapsed Gibbs sampler for the topics-over-time model
}
