% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{keyATMvb_call}
\alias{keyATMvb_call}
\title{Run the Variational Bayes for the keyATM models}
\usage{
keyATMvb_call(model)
}
\arguments{
\item{model}{A model}
}
\description{
Run the Variational Bayes for the keyATM models
}
\keyword{internal}
