% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyATMvb.R
\name{keyATMvb_fit}
\alias{keyATMvb_fit}
\title{Fit a keyATM model with Collapsed Variational Bayes}
\usage{
keyATMvb_fit(
  docs,
  model,
  no_keyword_topics,
  keywords = list(),
  model_settings = list(),
  vb_options = list(),
  priors = list(),
  options = list()
)
}
\description{
Fit a keyATM model with Collapsed Variational Bayes
}
\keyword{internal}
