% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lda_cov}
\alias{lda_cov}
\title{Run the Collapsed Gibbs sampler for LDA Dir-Multi (Mimno and McCalum 2008)}
\usage{
lda_cov(model, K, iter = 0L, output_iter = 10L)
}
\arguments{
\item{model}{A model, from \code{init} or a previous invocation of \code{train}, including a covariate}

\item{iter}{Required number of iterations}

\item{output_per}{Show log-likelihood and perplexity per this number during the iteration}
}
\description{
Run the Collapsed Gibbs sampler for LDA Dir-Multi (Mimno and McCalum 2008)
}
