% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LDA_weight}
\alias{LDA_weight}
\title{Run the Collapsed Gibbs sampler for LDA with weights}
\usage{
LDA_weight(model, iter = 0L, output_per = 10L)
}
\arguments{
\item{model}{A initialized model}

\item{iter}{Required number of iterations}

\item{output_per}{Show log-likelihood and perplexity per this number during the iteration}
}
\description{
Run the Collapsed Gibbs sampler for LDA with weights
}
