% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LDA_weight_tot}
\alias{LDA_weight_tot}
\title{Run the Collapsed Gibbs sampler for LDA topic-over-time with weights}
\usage{
LDA_weight_tot(model, iter = 0L, output_per = 10L, use_weight = 1L)
}
\arguments{
\item{model}{A model, from \code{init} or a previous invocation of \code{train}}

\item{iter}{Required number of iterations}

\item{output_per}{Show log-likelihood and perplexity per this number during the iteration}
}
\description{
Run the Collapsed Gibbs sampler for LDA topic-over-time with weights
}
