% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports (Eshima Shusei's conflicted copy
%   2023-05-07).R, R/RcppExports.R
\name{make_wsz_cpp}
\alias{make_wsz_cpp}
\title{Initialize assignments}
\usage{
make_wsz_cpp(docs_, info_, initialized_)

make_wsz_cpp(docs_, info_, initialized_)
}
\arguments{
\item{info}{Various information}

\item{initialized}{Store initialized objects (W, S and Z)}
}
\description{
Initialize assignments

Initialize assignments
}
\keyword{internal}
