% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiPGreg.R
\name{multiPGreg}
\alias{multiPGreg}
\title{Run multinomial regression with Polya-Gamma augmentation}
\usage{
multiPGreg(Y, X, num_topics, PG_params, iter = 1, store_lambda = 0)
}
\arguments{
\item{Y}{Outcomes.}

\item{X}{Covariates.}

\item{num_topics}{Number of topics.}

\item{PG_params}{Parameters used in this function.}

\item{iter}{The default is \code{1}.}

\item{store_lambda}{The default is \code{0}.}
}
\description{
Run multinomial regression with Polya-Gamma augmentation. There is no need to call this function directly. The keyATM Covariate internally uses this.
}
