% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{plot_alpha}
\alias{plot_alpha}
\title{Show a diagnosis plot of alpha}
\usage{
plot_alpha(x, start = 0, show_topic = NULL, thinning = 5,
  scale = "fixed")
}
\arguments{
\item{x}{the output from a keyATM model (see \code{keyATM()})}

\item{start}{Slice iteration}

\item{show_topic}{a vector to specify topic indexes to show}

\item{thinning}{a integer for thinning}

\item{scale}{a parameter to control the scale of y-axis: 'free' adjusts y-axis for parameters}
}
\value{
ggplot2 object
}
\description{
Show a diagnosis plot of alpha
}
