% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_p}
\alias{plot_p}
\title{Show a diagnosis plot of p}
\usage{
plot_p(x, show_topic = NULL)
}
\arguments{
\item{x}{the output from a keyATM model (see \code{keyATM()})}

\item{show_topic}{A vector to indicate topics to visualize}
}
\value{
ggplot2 object
}
\description{
Show a diagnosis plot of p
}
