% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_pi}
\alias{plot_pi}
\title{Show a diagnosis plot of pi}
\usage{
plot_pi(x, show_topic = NULL, start = 0)
}
\arguments{
\item{x}{the output from a keyATM model (see \code{keyATM()})}

\item{show_topic}{A vector to indicate topics to visualize}

\item{start}{slice iteration}
}
\value{
ggplot2 object
}
\description{
Show a diagnosis plot of pi
}
