% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_topicprop}
\alias{plot_topicprop}
\title{Show the expected proportion of the corpus belonging to each topic}
\usage{
plot_topicprop(x, show_topic = NULL, xmax = NULL, show_topwords = TRUE)
}
\arguments{
\item{x}{the output from a keyATM model (see \code{\link[=keyATM]{keyATM()}}).}

\item{show_topic}{an integer or a vector. Indicate topics to visualize. Default is \code{NULL}.}

\item{xmax}{a numeric. Indicate the max value on the x axis}

\item{show_topwords}{logical. Show topwords. The default is \code{TRUE}.}
}
\value{
keyATM_fig object
}
\description{
Show the expected proportion of the corpus belonging to each topic
}
