% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{posterior}
\alias{posterior}
\title{Get posterior quantities from model output}
\usage{
posterior(x, docnames = NULL)
}
\arguments{
\item{docnames}{Names of documents (not stored in the C++
seeededlda yet)}

\item{model_results}{The result of running \code{seededlda}}
}
\value{
a list with elements
  \itemize{
    \item{topics}{An N x K matrix of topic proportions}
    \item{terms}{A K x V matrix of topic-specific term generation
                 probabilities}
    \item{doclens}{A vector of document lengths in words, with which
                   Z can be reinflated from theta}
  }
}
\description{
Get posterior quantities from model output
}
