% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{posterior_tau}
\alias{posterior_tau}
\title{Calculate posterior tau}
\usage{
posterior_tau(res, topic_id = 1, cov_id = 1)
}
\arguments{
\item{topic_id}{}

\item{cov_id}{}

\item{x}{The result object from a fitted model}
}
\value{
A list that contains estimated tau for each iteration
}
\description{
Calculate posterior tau
}
