% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{posterior_theta}
\alias{posterior_theta}
\title{Calculate posterior theta}
\usage{
posterior_theta(res)
}
\arguments{
\item{x}{The result object from a fitted model}
}
\value{
A list that contains estimated theta for each iteration
}
\description{
Calculate posterior theta
}
