% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess_dictionary}
\alias{preprocess_dictionary}
\title{Preprocess a dictionary to make it usable as seed words}
\usage{
preprocess_dictionary(dict, corpus, min.freq = 5, ...)
}
\arguments{
\item{dict}{A quanteda dictionary, or the filename of something quanteda can
read in as a dictionary, or a named list which can be treated as
a dictionary}

\item{corpus}{A quanteda corpus}

\item{min.freq}{The fewest times a word can appear and remain in the dictionary}

\item{...}{Extra arguments to give to \code{dfm}}
}
\value{
A dictionary with entries filtered by frequency
}
\description{
This function take a dictionary, collapses to the top layer of
categories, adds all the corpus matches for any wildcarded pattern,
and removes anything that does not occur more than \code{min.freq}
times. The result is a dictionary suitable for use as seeds in
\code{topicdict}.
}
\examples{
dictfile <- system.file("extdata/laver-garry-ajps.ykd", package="topicdict")
# Use just the economics subtree of this dictionary
dict <- dictionary(dictfile)[['Laver and Garry']][["State in Economy"]]
data("corpus_uk_platforms")
new_dict <- choose_seeds(dict, corpus_uk_platforms, min.freq = 5)
}
