% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{read_dfm_cpp}
\alias{read_dfm_cpp}
\title{Read files from the quanteda dfm (this is the same as dgCMatrix)}
\usage{
read_dfm_cpp(dfm, W_read, vocab, show_progress_bar, split)
}
\arguments{
\item{dfm}{a dfm input (sparse Matrix)}

\item{vocab}{a vector of vocabulary}

\item{show_progress_bar}{show a progress bar}

\item{split}{split proportion}

\item{W_raw}{an object to return}
}
\description{
Read files from the quanteda dfm (this is the same as dgCMatrix)
}
\keyword{internal}
