% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topicdict_cpp.R
\name{read_seed_file}
\alias{read_seed_file}
\title{Read Seed file}
\usage{
read_seed_file(file_path)
}
\arguments{
\item{file_path}{a full path to the seed file (csv)}
}
\value{
a vector of seeds
}
\description{
Read Seed file
}
