% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{save.keyATM_output}
\alias{save.keyATM_output}
\title{Save a keyATM_output object}
\usage{
save.keyATM_output(x, file = stop("'file' must be specified"))
}
\arguments{
\item{x}{a keyATM_output object (see \code{keyATM()})}

\item{file}{a character}
}
\description{
Save a keyATM_output object
}
