% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{save.keyATM_viz}
\alias{save.keyATM_viz}
\title{Save a keyATM_viz object}
\usage{
save.keyATM_viz(x, file = stop("'file' must be specified"))
}
\arguments{
\item{x}{a keyATM_viz object (see \code{visualize_keywords()})}

\item{file}{a character}
}
\description{
Save a keyATM_viz object
}
