% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{save_fig.keyATM_viz}
\alias{save_fig.keyATM_viz}
\title{Save a keyword plot}
\usage{
save_fig.keyATM_viz(x, file = stop("'file' must be specified"))
}
\arguments{
\item{x}{a keyATM_viz object (see \code{visualize_keywords()})}

\item{file}{a character}
}
\description{
Save a keyword plot
}
