% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata_functions.R
\name{Save_Index_Freq}
\alias{Save_Index_Freq}
\title{Save index freq}
\usage{
Save_Index_Freq(doc_list, show_num, saveDir)
}
\arguments{
\item{doc_list}{document list}

\item{show_num}{how many to show}

\item{saveDir}{save directory}
}
\description{
Save index freq
}
