% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata_functions.R
\name{Save_seeds_len}
\alias{Save_seeds_len}
\title{Save index freq}
\usage{
Save_seeds_len(doc_list, saveDir)
}
\arguments{
\item{doc_list}{document list}

\item{saveDir}{save directory}
}
\description{
Save index freq
}
