% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topicdict_cpp.R
\name{seededlda}
\alias{seededlda}
\title{Run Seeded LDA}
\usage{
seededlda(data_folder_path, seed_csv_path, num_regular_topic = 0,
  iter_num = 10, show_words_num = 10, full_output = F, rand_seed = 0)
}
\arguments{
\item{data_folder_path}{Full path to the data folder.}

\item{seed_csv_path}{Full path to seed.}

\item{num_regular_topic}{a number of regular topic}

\item{iter_num}{iteration number.}

\item{show_words_num}{Number of words to show in the output object.}

\item{full_output}{Return topic assignments and indicators for all words.}

\item{seed}{seed for random number generators (not working well??).}
}
\value{
Seeded LDA results.
}
\description{
Run Seeded LDA
}
\examples{
\dontrun{seededlda(data_folder_path="data/is/stored/here/", seed_path="path/to/seed.txt")}

}
