% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{semantic_coherence}
\alias{semantic_coherence}
\title{Semantic Coherence: Mimno et al. (2011)}
\usage{
semantic_coherence(x, docs, n = 10)
}
\arguments{
\item{x}{the output from a keyATM model (see \code{\link[=keyATM]{keyATM()}}).}

\item{docs}{texts read via \code{\link[=keyATM_read]{keyATM_read()}}.}

\item{n}{integer. The number terms to visualize. Default is \code{10}.}
}
\value{
A vector of topic coherence metric calculated by each topic.
}
\description{
Mimno, David et al. 2011. “Optimizing Semantic Coherence in Topic Models.”
In Proceedings of the 2011 Conference on Empirical Methods in Natural
Language Processing, Edinburgh, Scotland, UK.: Association for Computational
Linguistics, 262–72. https://aclanthology.org/D11-1024.
}
\details{
Equation 1 of Mimno et al. 2011 adopted to keyATM.
}
