% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{set_doc_names}
\alias{set_doc_names}
\title{Set document names}
\usage{
set_doc_names(x, doc_names)
}
\arguments{
\item{x}{the output from a keyATM model (see \code{keyATM_output})}

\item{doc_names}{new names for documents}
}
\value{
an output object with new document names in its components
}
\description{
Set document names
}
