% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{top_terms}
\alias{top_terms}
\title{Show the top terms for each topic}
\usage{
top_terms(x, n = 10, measure = c("probability", "lift"))
}
\arguments{
\item{x}{The posterior from a fitted model (see \code{posterior})}

\item{n}{How many terms to show: default NULL shows all}

\item{measure}{How to sort the terms: 'probability' (default) or 'lift'}
}
\value{
An n x k table of the top n words in each topic
}
\description{
Show the top terms for each topic
}
