% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{top_topics}
\alias{top_topics}
\title{Show the top topics for each document}
\usage{
top_topics(x, n = 2, measure = c("probability", "lift"))
}
\arguments{
\item{x}{the output from a keyATM model (see \code{keyATM_output})}

\item{n}{How many topics to show. Default: 2}

\item{measure}{How to sort the topics: 'probability' (default) or 'lift'}
}
\value{
An n x k table of the top n topics in each document
}
\description{
Show the top topics for each document
}
