% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topicdict-package.R
\docType{package}
\name{topicdict}
\alias{topicdict}
\alias{topicdict-package}
\title{topicdict: Topic Modeling Informed by a Content Analysis Dictionary}
\description{
The \code{\link{topicdict_model}} function initializes a model from a
seed dictionary and
a file pattern specifying documents.  The \code{\link{topicdict_train}}
function runs a Gibbs sampler
on the this structure. The \code{\link{posterior}} function generates
useful summary statistics, e.g. theta and beta matrices.
}
