% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{topicdict_train}
\alias{topicdict_train}
\title{Run the Gibbs sampler}
\usage{
topicdict_train(model, alpha_k, iter = 0L)
}
\arguments{
\item{model}{A model, from \code{init} or a previous invocation of \code{train}}

\item{alpha_k}{A starting value for alpha (will be removed when alpha updates are back in)}

\item{iter}{Required number of iterations}
}
\description{
Run the Gibbs sampler
}
