% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{topicdict_train_totcov}
\alias{topicdict_train_totcov}
\title{Run the Collapsed Gibbs sampler for the topic-over-time model with covariate}
\usage{
topicdict_train_totcov(model, iter = 0L, output_per = 10L)
}
\arguments{
\item{model}{A model, from \code{init} or a previous invocation of \code{train}}

\item{iter}{Required number of iterations}

\item{output_per}{Show log-likelihood and perplexity per this number during the iteration}
}
\description{
Run the Collapsed Gibbs sampler for the topic-over-time model with covariate
}
