% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{train}
\alias{train}
\title{Run the Gibbs sampler}
\usage{
train(model, k_seeded, k_free, alpha_k, iter = 0L)
}
\arguments{
\item{model}{A model, from \code{init} or a previous invocation of \code{train}}

\item{k_seeded}{How many topics are seeded}

\item{k_free}{How many regular unseeded topics are required}

\item{alpha_k}{A starting value for alpha (will be removed when alpha updates are back in)}

\item{iter}{Required number of iterations}
}
\description{
Run the Gibbs sampler
}
