% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{values_fig}
\alias{values_fig}
\title{Get values used to create a figure}
\usage{
values_fig(x)
}
\arguments{
\item{x}{the keyATM_fig object.}
}
\description{
Get values used to create a figure
}
\seealso{
\code{\link[=save_fig]{save_fig()}}, \code{\link[=visualize_keywords]{visualize_keywords()}}, \code{\link[=plot_alpha]{plot_alpha()}}, \code{\link[=plot_modelfit]{plot_modelfit()}}, \code{\link[=plot_pi]{plot_pi()}}, \code{\link[=plot_timetrend]{plot_timetrend()}}, \code{\link[=by_strata_DocTopic]{by_strata_DocTopic()}}
}
