% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyATM.R
\name{weightedLDA}
\alias{weightedLDA}
\title{Weighted LDA Main function}
\usage{
weightedLDA(docs, model, number_of_topics, model_settings = list(),
  priors = list(), options = list(), keep = c())
}
\arguments{
\item{docs}{texts read via \code{keyATM_read()}}

\item{model}{Weighted LDA model: "base", "covariates", and "dynamic"}

\item{number_of_topics}{the number of regular topics}

\item{model_settings}{a list of model specific settings}

\item{priors}{a list of priors of parameters}

\item{options}{a list of options}

\item{keep}{a vector of the names of elements you want to keep in output}
}
\value{
A keyATM_output object containing:
  \describe{
    \item{number_of_topics}{Number of regular unseeded topics}
    \item{V}{Number of word types}
    \item{N}{Number of documents}
    \item{theta}{Normalized topic proportions for each document}
    \item{phi}{Normalized topic specific word generation probabilities}
    \item{topic_counts}{Number of tokens assigned to each topic}
    \item{word_counts}{Number of times each word type appears}
    \item{doc_lens}{Length of each document in tokens}
    \item{vocab}{Words in the vocabulary}
    \item{model_fit}{Perplexity and log-likelihood}
    \item{p}{Estimated p}
    \item{values_iter}{Organized values stored during iterations}
    \item{kept_values}{Outputs you specified to store.}
  }
}
\description{
Run weighted LDA models.
}
\examples{
\dontrun{
  # Weighted LDA
  out <- weightedLDA(
                     keyATM_docs, model = "base", number_of_topics = 5
                    )

  # Weighted LDA Covariates
  out <- weightedLDA(
                     keyATM_docs, model = "covariates", number_of_topics = 5,
                     model_settings(covariates_data = cov_matrix)
                    )                   

  # Weighted LDA Dynamic
  out <- weightedLDA(
                     keyATM_docs, model = "dynamic", number_of_topics = 5,
                     model_settings(time_index = time_index_vec, num_states = 5)
                    )

}

}
