% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports (Eshima Shusei's conflicted copy
%   2023-05-07).R, R/RcppExports.R
\name{word_in_doc}
\alias{word_in_doc}
\title{Checking if a word is in a document}
\usage{
word_in_doc(doc, word)

word_in_doc(doc, word)
}
\arguments{
\item{doc}{a vector}

\item{word}{a word to check}
}
\value{
bool

bool
}
\description{
Checking if a word is in a document

Checking if a word is in a document
}
\keyword{internal}
