% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{by_strata_DocTopic}
\alias{by_strata_DocTopic}
\title{Estimate document-topic distribution by strata (for covariate models)}
\usage{
by_strata_DocTopic(
  x,
  by_name,
  by_values,
  burn_in = NULL,
  parallel = TRUE,
  mc.cores = NULL,
  posterior_mean = FALSE
)
}
\arguments{
\item{x}{the output from a keyATM model (see \code{keyATM()})}

\item{by_name}{character. The name of the variable to use.}

\item{by_values}{numeric. The values of the variable specified in `by_name`}

\item{burn_in}{integer. Burn-in period. If not specified, it is the half of samples. Default is \code{NULL}.}

\item{parallel}{logical. If \code{TRUE}, parallelization for speeding up. Default is \code{TRUE}.}

\item{mc.cores}{integer. The number of cores to use. Default is \code{NULL}.}

\item{posterior_mean}{logical. If \code{TRUE}, the quantity of interest to estimate is the posterior mean. Default is \code{FALSE}.}
}
\value{
strata_topicword object (a list)
}
\description{
Estimate document-topic distribution by strata (for covariate models)
}
