% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_PGtheta_R}
\alias{calc_PGtheta_R}
\title{Calculate the probability for Polya-Gamma Covariate Model}
\usage{
calc_PGtheta_R(theta_tilda, theta, num_doc, num_topics)
}
\arguments{
\item{model}{A initialized model}

\item{iter}{Required number of iterations}
}
\description{
Same as utils::calc_PGtheta, but this is for calling from R
}
\keyword{internal}
