% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{covariates_info}
\alias{covariates_info}
\title{Show covariates information}
\usage{
covariates_info(x)
}
\arguments{
\item{x}{the output from the covariate keyATM model (see \code{\link[=keyATM]{keyATM()}}).}
}
\description{
Show covariates information
}
