% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata_call.R
\name{create_sim_data}
\alias{create_sim_data}
\title{Create Simulation Data}
\usage{
create_sim_data(saveDir, D = 200, K = 10, TotalV = 8000, alpha = 0.1,
  beta_r = 0.1, beta_s = 0.1, p = NULL, gamma1 = NULL, gamma2 = NULL,
  lambda = 300, seeds_len = 5, prop_seeds = NULL, rand_seed = 123)
}
\arguments{
\item{saveDir}{data storage (set full path)}

\item{D}{the number of documents}

\item{K}{the number of topics}

\item{TotalV}{maximum number of unique words}

\item{alpha}{a parameter of dirichlet of theta, a vector or a scalar}

\item{beta_r}{a parameter of dirichlet of phi_regular}

\item{beta_s}{a parameter of dirichlet of phi_seed}

\item{p}{a probability vector of using seed (set either gamma or p)}

\item{gamma1}{a parameter of beta (set either gamma or p)}

\item{gamma2}{a parameter of beta (set either gamma or p)}

\item{lambda}{a parameter of poisson (average length of documents)}

\item{seeds_len}{the number of seeds words per each topic}

\item{prop_seeds}{the proportion of seed words}
}
\description{
Create Simulation Data
}
\examples{
create_sim_data(saveDir="/Users/Shusei/Desktop/temp/SimulationData/", D=200, K=10, TotalV=8000, alpha=0.1, beta_r=0.1, beta_s=0.1, gamma1=1, gamma2=20, lambda=300, seeds_len=5)
create_sim_data(saveDir="/Users/Shusei/Desktop/temp/SimulationData/", D=200, K=10, TotalV=8000, alpha=0.1, beta_r=0.1, beta_s=0.1, p=rep(0.2, 10), lambda=300, seeds_len=5)

}
