% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{diagnosis_model_fit}
\alias{diagnosis_model_fit}
\title{Show a diagnosis plot of log-likelihood and perplexity}
\usage{
diagnosis_model_fit(x, start = 1)
}
\arguments{
\item{x}{the output from a keyATM model (see \code{keyATM_output()})}

\item{start}{}
}
\value{
ggplot2 object
}
\description{
Show a diagnosis plot of log-likelihood and perplexity
}
