% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{diagnosis_model_fit}
\alias{diagnosis_model_fit}
\title{Show a diagnosis plot of log-likelihood and perplexity}
\usage{
diagnosis_model_fit(x, start = NULL)
}
\arguments{
\item{x}{The posterior from a fitted model (see \code{posterior})}

\item{start}{Slice iteration}
}
\value{
ggplot2 object
}
\description{
Show a diagnosis plot of log-likelihood and perplexity
}
