% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{diagnosis_p}
\alias{diagnosis_p}
\title{Show a diagnosis plot of p}
\usage{
diagnosis_p(x, topicvec = c())
}
\arguments{
\item{x}{The posterior from a fitted model (see \code{posterior})}

\item{topicvec}{A topic vector to reorder}
}
\value{
ggplot2 object
}
\description{
Show a diagnosis plot of p
}
