% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyATM-package.R
\docType{package}
\name{keyATM-package}
\alias{keyATM}
\alias{keyATM-package}
\title{keyATM: Topic modeling with predetermined keywords}
\description{
Fits a topic model using keywords.
}
\author{
\strong{Maintainer}: Shusei Eshima \email{shuseieshima@g.harvard.edu}

Authors:
\itemize{
  \item Tomoya Sasaki \email{tomoyas@mit.edu}
  \item William Lowe \email{wlowe@princeton.edu}
  \item Kosuke Imai \email{imai@harvard.edu}
}

}
\keyword{internal}
