% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyATM-package.R
\docType{package}
\name{keyATM-package}
\alias{keyATM-package}
\title{Keyword Assisted Topic Models}
\description{
The implementation of keyATM models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://keyatm.github.io/keyATM/}
  \item Report bugs at \url{https://github.com/keyATM/keyATM/issues}
}

}
\author{
\strong{Maintainer}: Shusei Eshima \email{shuseieshima@g.harvard.edu} (\href{https://orcid.org/0000-0003-3613-4046}{ORCID})

Authors:
\itemize{
  \item Tomoya Sasaki \email{tomoyas@mit.edu}
  \item Kosuke Imai \email{imai@harvard.edu}
}

Other contributors:
\itemize{
  \item William Lowe \email{wlowe@princeton.edu} [contributor]
  \item Chung-hong Chan \email{chainsawtiney@gmail.com} (\href{https://orcid.org/0000-0002-6232-7530}{ORCID}) [contributor]
  \item Romain François (\href{https://orcid.org/0000-0002-2444-4226}{ORCID}) [contributor]
}

}
