% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyATM.R
\name{keyATM}
\alias{keyATM}
\title{keyATM main function}
\usage{
keyATM(docs, model, no_keyword_topics, keywords = list(),
  model_settings = list(), priors = list(), options = list(),
  keep = c())
}
\arguments{
\item{docs}{texts read via \code{keyATM_read()}}

\item{model}{keyATM model: "base", "covariates", and "dynamic"}

\item{no_keyword_topics}{the number of regular topics}

\item{keywords}{a list of keywords}

\item{model_settings}{a list of model specific settings}

\item{priors}{a list of priors of parameters}

\item{options}{a list of options 
\describe{
     \item{seed}{A numeric value for random seed. If it is not provided, the package randomly selects a seed.}
     \item{iterations}{An integer. Number of iterations. Default is 1500.}
     \item{verbose}{If `TRUE` print loglikelihood and perplexity. Default is `FALSE`.}
     \item{llk_per}{An integer. If the value is `j` **keyATM** stores loglikelihood and perplexity every `j` iteration. Default value is 10 per iterations}
     \item{use_weights}{If `TRUE` use weight. Default is `TRUE`.}
     \item{prune}{If `TRUE` rume keywords that do not appear in the corpus. Default is `TRUE`.}
     \item{thinning}{An integer. If the value is `j` **keyATM** stores following parameters every `j` iteration. \itemize{
           \item{theta}{For all models. If `store_theta` is `TRUE` document-level topic assignment is stored (sufficient statistics to calculate document-topic distributions `theta`).}
           \item{alpha}{For the base and dynamic models. In the base model alpha is shared across all documents whereas each state has different alpha in the dynamic model.}
           \item{lambda}{For the covariate model.}
           \item{R}{For the dynamic model. The state each document belongs to.}
           \item{P}{For the dynamic model. The state transition probability}
           }
     }
}}

\item{keep}{a vector of the names of elements you want to keep in output}
}
\value{
A keyATM_output object containing:
  \describe{
    \item{keyword_k}{number of keyword topics}
    \item{no_keyword_topics}{number of regular unseeded topics}
    \item{V}{number of word types}
    \item{N}{number of documents}
    \item{theta}{normalized topic proportions for each document}
    \item{phi}{normalized topic specific word generation probabilities}
    \item{topic_counts}{number of tokens assigned to each topic}
    \item{word_counts}{number of times each word type appears}
    \item{doc_lens}{length of each document in tokens}
    \item{vocab}{words in the vocabulary}
    \item{model_fit}{perplexity and log-likelihood}
    \item{p}{estimated p}
    \item{values_iter}{organized values stored during iterations}
    \item{kept_values}{outputs you specified to store}
  }
}
\description{
Run keyATM models.
}
\examples{
\dontrun{
  library(keyATM)
  library(quanteda)
  data(keyATM_data_bills)
  bills_keywords <- keyATM_data_bills$keywords
  bills_dfm <- keyATM_data_bills$doc_dfm  # quanteda dfm object
  keyATM_docs <- keyATM_read(bills_dfm)
  # keyATM Base
  out <- keyATM(
                docs, model = "base", no_keyword_topics = 5, keywords = keywords_list
               )

  # keyATM Covariates
  out <- keyATM(
                docs, model = "covariates", no_keyword_topics = 5, keywords = keywords_list,
                model_settings(covariates_data = cov)
               )

  # keyATM Dynamic
  out <- keyATM(
                docs, model = "dynamic", no_keyword_topics = 5, keywords = keywords_list,
                model_settings(time_index = time_index_vec, num_states = 5)
               )

}

}
\seealso{
\url{https://keyatm.github.io/keyATM/articles/pkgdown_files/Options.html}
}
