% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{keyATM_fit_cov}
\alias{keyATM_fit_cov}
\title{Run the Collapsed Gibbs sampler for the keyATM covariates}
\usage{
keyATM_fit_cov(model, iter = 0L)
}
\arguments{
\item{model}{A initialized model}

\item{iter}{Required number of iterations}
}
\description{
Run the Collapsed Gibbs sampler for the keyATM covariates
}
\keyword{internal}
