% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{keyATM_fit_covPG}
\alias{keyATM_fit_covPG}
\title{Run the Collapsed Gibbs sampler for the keyATM covariates (Polya-Gamma)}
\usage{
keyATM_fit_covPG(model, iter = 0L)
}
\arguments{
\item{model}{A initialized model}

\item{iter}{Required number of iterations}
}
\description{
Run the Collapsed Gibbs sampler for the keyATM covariates (Polya-Gamma)
}
\keyword{internal}
