% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{keyATM_fit_LDA}
\alias{keyATM_fit_LDA}
\title{Run the Collapsed Gibbs sampler for weighted LDA}
\usage{
keyATM_fit_LDA(model, iter = 0L)
}
\arguments{
\item{model}{A initialized model}

\item{iter}{Required number of iterations}
}
\description{
Run the Collapsed Gibbs sampler for weighted LDA
}
\keyword{internal}
