% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{keyATM_fit_LDAcov}
\alias{keyATM_fit_LDAcov}
\title{Run the Collapsed Gibbs sampler for weighted LDA with covariates}
\usage{
keyATM_fit_LDAcov(model, resume = FALSE)
}
\arguments{
\item{model}{A initialized model}

\item{resume}{resume or not}
}
\description{
Run the Collapsed Gibbs sampler for weighted LDA with covariates
}
\keyword{internal}
