% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{keyATM_fit_LDAHMM}
\alias{keyATM_fit_LDAHMM}
\title{Run the Collapsed Gibbs sampler for the weighted LDA with HMM model}
\usage{
keyATM_fit_LDAHMM(model, resume = FALSE)
}
\arguments{
\item{model}{A initialized model}

\item{resume}{resume or not}
}
\description{
Run the Collapsed Gibbs sampler for the weighted LDA with HMM model
}
\keyword{internal}
