% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{keyATM_output}
\alias{keyATM_output}
\title{Get posterior quantities from model output}
\usage{
keyATM_output(model)
}
\arguments{
\item{model}{a fitted keyATM model}
}
\value{
A list containing:
  \describe{
    \item{seed_K}{ Number of keyword topics}
    \item{regular_k}{ Number of regular unseeded topics}
    \item{V}{ Number of word types}
    \item{N}{ Number of documents}
    \item{theta}{ Normalized tpoic proportions for each document}
    \item{beta}{ Normalized topic specific word generation probabilities}
    \item{topic_counts}{ Number of tokens assigned to each topic}
    \item{word_counts}{ Number of times each word type appears}
    \item{doc_lens}{ Length of each document in tokens}
    \item{vocab}{ Words in the vocabulary}
    \item{alpha}{ \code{alpha} during the iteration}
    \item{modelfit}{ Perplexity and log-likelihood}
    \item{p}{ Estimated p}
    \item{options}{ Options used in the \code{model}}
  }
}
\description{
Constructs a (N x K) matrix \code{theta} and (K x V) matrix \code{beta}
plus their margins from the sample of Z and W in \code{model}.
\code{model} is an output of \code{keyATM_run()}.
These statistics implicitly marginalize over X.
}
