% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{keyATM_output}
\alias{keyATM_output}
\title{\code{keyATM_fit()} calls \code{keyATM_output()}}
\usage{
keyATM_output(model)
}
\description{
\code{keyATM_fit()} calls \code{keyATM_output()}
}
\keyword{internal}
