% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{keyATM_read}
\alias{keyATM_read}
\title{Read Texts}
\usage{
keyATM_read(texts, encoding = "UTF-8", check = TRUE)
}
\arguments{
\item{texts}{Input. keyATM takes dfm, data.frame, tibble, and a vector of file paths.}
}
\value{
A list whose elements are splitted texts.
}
\description{
This function read texts and creates a tibble.
}
\examples{
\dontrun{
 # Use quanteda dfm 
 keyATM_docs <- keyATM_read(quanteda_dfm) 
  
 # Use data.frame or tibble (texts should be stored in a column named `text`)
 keyATM_docs <- keyATM_read(data_frame_object) 
 keyATM_docs <- keyATM_read(tibble_object) 

 # Use a vector that stores full paths to the files  
 files <- list.files(doc_folder, pattern = "*.txt", full.names = T) 
 keyATM_docs <- keyATM_read(files) 
}

}
