% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{make_wsz_cpp}
\alias{make_wsz_cpp}
\title{Initialize assignments}
\usage{
make_wsz_cpp(docs_, info_, initialized_)
}
\arguments{
\item{docs_}{Documents}

\item{info_}{Various information}

\item{initialized_}{Store initialized objects (W, S and Z)}
}
\description{
Initialize assignments
}
\keyword{internal}
