% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{plot.strata_doctopic}
\alias{plot.strata_doctopic}
\title{Plot Document-Topic distribution by strata}
\usage{
\method{plot}{strata_doctopic}(x, topics = NULL, quantile_vec = c(0.05,
  0.5, 0.95), ...)
}
\arguments{
\item{x}{a strata_doctopic object (see \code{by_strata_DocTopic()})}

\item{topics}{topics to show}

\item{quantile_vec}{quantiles to show}
}
\value{
ggplot2 object
}
\description{
Plot Document-Topic distribution by strata
}
