% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.strata_doctopic}
\alias{plot.strata_doctopic}
\title{Plot document-topic distribution by strata (for covariate models)}
\usage{
\method{plot}{strata_doctopic}(x, topics = NULL, quantile_vec = c(0.05, 0.5, 0.95), ...)
}
\arguments{
\item{x}{a strata_doctopic object (see \code{by_strata_DocTopic()})}

\item{topics}{a vector or an integer. Indicate topics to visualize.}

\item{quantile_vec}{a numeric. Quantiles to visualize}

\item{...}{additional arguments not used}
}
\value{
ggplot2 object
}
\description{
Plot document-topic distribution by strata (for covariate models)
}
