% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_alpha}
\alias{plot_alpha}
\title{Show a diagnosis plot of alpha}
\usage{
plot_alpha(x, start = 0, show_topic = NULL, scales = "fixed")
}
\arguments{
\item{x}{the output from a keyATM model (see \code{\link[=keyATM]{keyATM()}}).}

\item{start}{integer. The start of slice iteration. Default is \code{0}.}

\item{show_topic}{a vector to specify topic indexes to show. Default is \code{NULL}.}

\item{scales}{character. Control the scale of y-axis (the parameter in \link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}): \code{free} adjusts y-axis for parameters. Default is \code{fixed}.}
}
\value{
keyATM_fig object
}
\description{
Show a diagnosis plot of alpha
}
\seealso{
\code{\link[=save_fig]{save_fig()}}
}
