% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_modelfit}
\alias{plot_modelfit}
\title{Show a diagnosis plot of log-likelihood and perplexity}
\usage{
plot_modelfit(x, start = 1)
}
\arguments{
\item{x}{the output from a keyATM model (see \code{\link[=keyATM]{keyATM()}}).}

\item{start}{integer. The starting value of iteration to use in plot. Default is \code{1}.}
}
\value{
keyATM_fig object.
}
\description{
Show a diagnosis plot of log-likelihood and perplexity
}
\seealso{
\code{\link[=save_fig]{save_fig()}}
}
