% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_pi}
\alias{plot_pi}
\title{Show a diagnosis plot of pi}
\usage{
plot_pi(x, show_topic = NULL, start = 0)
}
\arguments{
\item{x}{the output from a keyATM model (see \code{\link[=keyATM]{keyATM()}})}

\item{show_topic}{an integer or a vector. Indicate topics to visualize. Default is \code{NULL}.}

\item{start}{integer. The starting value of iteration to use in the plot. Default is \code{0}.}
}
\value{
keyATM_fig object
}
\description{
Show a diagnosis plot of pi
}
\seealso{
\code{\link[=save_fig]{save_fig()}}
}
