% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_pi}
\alias{plot_pi}
\title{Show a diagnosis plot of pi}
\usage{
plot_pi(x, show_topic = NULL, start = 0)
}
\arguments{
\item{x}{the output from a keyATM model (see \code{keyATM()})}

\item{show_topic}{an integer or a vector. Indicate topics to visualize. Default is \code{NULL}.}

\item{start}{integer. The starting value of iteration to use in the plot. Default is 0.}
}
\value{
ggplot2 object
}
\description{
Show a diagnosis plot of pi
}
