% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_topicprop}
\alias{plot_topicprop}
\title{Show the expected proportion of the corpus belonging to each topic}
\usage{
plot_topicprop(
  x,
  n = 3,
  show_topic = NULL,
  show_topwords = TRUE,
  label_topic = NULL,
  order = c("proportion", "topicid"),
  xmax = NULL
)
}
\arguments{
\item{x}{the output from a keyATM model (see \code{\link[=keyATM]{keyATM()}}).}

\item{n}{The number of top words to show. Default is \code{3}.}

\item{show_topic}{an integer or a vector. Indicate topics to visualize. Default is \code{NULL}.}

\item{show_topwords}{logical. Show topwords. The default is \code{TRUE}.}

\item{label_topic}{a character vector. The name of the topics in the plot.}

\item{order}{The order of topics.}

\item{xmax}{a numeric. Indicate the max value on the x axis}
}
\value{
keyATM_fig object
}
\description{
Show the expected proportion of the corpus belonging to each topic
}
\seealso{
\code{\link[=save_fig]{save_fig()}}
}
