% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{save_fig}
\alias{save_fig}
\title{Save a figure}
\usage{
save_fig(x, filename, ...)
}
\arguments{
\item{x}{the object}

\item{filename}{file name to create on disk}

\item{...}{other arguments passed on to the \link[ggplot2:ggsave]{ggplot2::ggsave()} function}
}
\description{
Save a figure
}
\seealso{
\code{\link[=visualize_keywords]{visualize_keywords()}}, \code{\link[=plot_alpha]{plot_alpha()}}, \code{\link[=plot_modelfit]{plot_modelfit()}}, \code{\link[=plot_pi]{plot_pi()}}, \code{\link[=by_strata_DocTopic]{by_strata_DocTopic()}}
}
