% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{save_fig}
\alias{save_fig}
\title{Save a figure}
\usage{
save_fig(x, filename, ...)
}
\arguments{
\item{x}{the object}

\item{filename}{file name to create on disk}

\item{...}{other arguments passed on to the \code{ggsave()} function}
}
\description{
Save a figure
}
\seealso{
\code{\link{visualize_keywords}}, \code{\link{plot_alpha}}, \code{\link{plot_modelfit}}, and \code{\link{plot_pi}}
}
