% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{set_topic_names}
\alias{set_topic_names}
\title{Set topic names}
\usage{
set_topic_names(x, topic_names)
}
\arguments{
\item{x}{the output from a keyATM model (see \code{keyATM_output})}

\item{topic_names}{new names for topics}
}
\value{
an output object with new topic names in its components
}
\description{
Set topic names
}
