% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{top_docs}
\alias{top_docs}
\title{Show the top documents for each topic}
\usage{
top_docs(x, n = 10, measure = c("probability", "lift"))
}
\arguments{
\item{x}{The posterior from a fitted model (see \code{posterior})}

\item{n}{How many documents to show: default 10}

\item{measure}{How to sort the terms: 'probability' (default) or 'lift'}
}
\value{
An n x k table of the top n documents for each topic
}
\description{
Show the top documents for each topic
}
