% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{top_docs}
\alias{top_docs}
\title{Show the top documents for each topic}
\usage{
top_docs(x, n = 10)
}
\arguments{
\item{x}{the output from a keyATM model (see \code{\link[=keyATM]{keyATM()}}).}

\item{n}{How many documents to show. Default is \code{10}.}
}
\value{
An n x k table of the top n documents for each topic, each number is a document index.
}
\description{
Show the top documents for each topic
}
