% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{top_terms}
\alias{top_terms}
\title{Show the top terms for each topic}
\usage{
top_terms(x, n = 10, measure = c("probability", "lift"),
  show_seed = TRUE)
}
\arguments{
\item{x}{The posterior from a fitted model (see \code{posterior})}

\item{n}{How many terms to show. Default: NULL, which shows all}

\item{measure}{How to sort the terms: 'probability' (default) or 'lift'}

\item{show_seed}{Mark seeded vocabulary. See below for details (Default: TRUE)}
}
\value{
An n x k table of the top n words in each topic
}
\description{
If \code{show_seed} is true then words in their seeded categories
are suffixed with a check mark. Words from another seeded category
are labeled with the name of that category.
}
