% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{top_topics}
\alias{top_topics}
\title{Show the top topics for each document}
\usage{
top_topics(x, measure = c("probability", "lift"), n = 2)
}
\arguments{
\item{x}{The posterior from a fitted model (see \code{posterior})}

\item{measure}{How to sort the topics: 'probability' (default) or 'lift'}

\item{n}{How many topics to show. Default: 2}
}
\value{
An n x k table of the top n topics in each document
}
\description{
Show the top topics for each document
}
