% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{top_words}
\alias{top_words}
\title{Show the top words for each topic}
\usage{
top_words(x, n = 10, measure = c("probability", "lift"), show_keyword = TRUE)
}
\arguments{
\item{x}{the output (see \code{\link[=keyATM]{keyATM()}} and \code{\link[=by_strata_TopicWord]{by_strata_TopicWord()}}).}

\item{n}{integer. The number terms to visualize. Default is \code{10}.}

\item{measure}{character. The way to sort the terms: \code{probability} (default) or \code{lift}.}

\item{show_keyword}{logical. If \code{TRUE}, mark keywords. Default is \code{TRUE}.}
}
\value{
An n x k table of the top n words in each topic
}
\description{
If \code{show_keyword} is \code{TRUE} then words in their keyword topics
are suffixed with a check mark. Words from another keyword topic
are labeled with the name of that category.
}
