% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topicdict-package.R
\docType{package}
\name{topicdict}
\alias{topicdict}
\alias{topicdict-package}
\title{topicdict: Topic Modeling Informed by a Content Analysis Dictionary}
\description{
The init function initializes a model from a seed dictionary and
a list of documents. The train function runs a Gibbs sampler
on the same data structure. The posterior function generates
useful summary statistics, e.g. theta and beta matrices.
}
