% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{topicdict_train}
\alias{topicdict_train}
\title{Run the Gibbs sampler}
\usage{
topicdict_train(model, iter = 0L, output_per = 10L, eta_1 = 1,
  eta_2 = 1, eta_1_regular = 2, eta_2_regular = 1)
}
\arguments{
\item{model}{A model, from \code{init} or a previous invocation of \code{train}}

\item{iter}{Required number of iterations}

\item{output_per}{Show log-likelihood and perplexity per this number during the iteration}
}
\description{
Run the Gibbs sampler
}
