% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{visualize_keywords}
\alias{visualize_keywords}
\title{Visualize keywords}
\usage{
visualize_keywords(docs, keywords, prune = TRUE, label_size = 3.2)
}
\arguments{
\item{docs}{A keyATM_docs-class object, generated by \code{keyATM_read()} function}

\item{keywords}{A list of keywords. The length}

\item{prune}{logical; If \code{TRUE}, prune keywords that do not appear in \code{docs}. Default is \code{TRUE}.}

\item{label_size}{Numeric value that specifies the size of the keyword labels appearing in the plot. Default is 3.2.}
}
\value{
A keyATM_viz-class object, which is a list containing: \describe{
   \item{figure}{a ggplot2 object}
   \item{values}{a tibble object that stores values used to create the plot}
   \item{keywords}{a list of keywords that appear in documents}
}
}
\description{
This function visualizes the proportion of keywords in the documents.
}
\examples{
\dontrun{
 # Prepare a keyATM_docs object
 keyATM_docs <- keyATM_read(input) 
  
 # Keywords are in a list with its length equals to the number of keyword topics
 keywords <- list(
                   c("education", "child", "student"),  # Education
                   c("public", "health", "program"),  # Health
                 )

 # Create an object to visualize keywords
 keyATM_viz <- visualize_keywords(keyATM_docs, keywords)

 # View a figure
 keyATM_viz
   # Or: `keyATM_viz$figure`

 # Save a figure 
 filename = "visualize_keywords.pdf"
 save_fig(keyATM_viz, filename)

}

}
